Mais Atributos de Aventureiro e correção de altura de Classe pessoa

+++ src/br/com/guilda/Pessoa.java
@@ -4,9 +4,9 @@ public class Pessoa {
     private String nome;
     private String genero;
     private int idade;
-    private int altura;
+    private double altura;
 
-    public Pessoa(String nome, String genero, int idade, int altura){
+    public Pessoa(String nome, String genero, int idade, double altura){
         this.nome = nome;
         this.genero = genero;
         this.idade = idade;
